+++ src/views/Sections/SectionProjects.js
@@ -2,20 +2,16 @@ import React from 'react';
 
 //import from material-ui core
 import {withStyles, ThemeProvider} from '@material-ui/core/styles';
-import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 //import from template componets
 import TitleGridContainer from 'views/Components/TitleGridContainer';
+import ListPanel from 'views/Components/ListPanel';
 
 //react-markdown
 import ReactMarkdown from 'react-markdown/with-html';
 
 //style and theme
 import projectStyle from 'assets/jss/material-kit-react/sections/projectsStyle';
-import panelTheme from 'assets/theme/expansionPanelTheme.js';
 
 //project data
 import {getProjects, projectDescription} from 'assets/data/projectsData';
@@ -34,53 +30,25 @@ class SectionProjects extends React.Component{
         .then(res=>this.setState({...this.state, md:res}))
     }
 
-    handlePanelExpand(panelName){
-        if(this.state.expand === panelName){
-            this.setState({...this.state, expand: ''})
-        }
-        else{
-            this.setState({...this.state, expand: panelName})
-        } 
-    }
-
-    renderProjectMD(projects){
-        const {classes} = this.props;
+    transformProjectToPanels(projects){
         if(!projects) return null;
-        return(
-            <ThemeProvider theme={panelTheme}>
+        const {classes} = this.props;
+        return projects.map(project=>{
+            return {
+                summary: (<div className={classes.panelTitle}>{project.title}</div>),
+                detail: (
                 <div>
-                {
-                    projects.map((exp, index)=>{
-                        return (
-                            <ExpansionPanel
-                            key={index}
-                            expanded={this.state.expand === `panel${index+1}`} 
-                            onChange={()=>this.handlePanelExpand(`panel${index+1}`)}>
-                                <ExpansionPanelSummary
-                                expandIcon={<ExpandMoreIcon />}
-                                >
-                                    <div className={classes.panelTitle}>{exp.title}</div>
-                                </ExpansionPanelSummary>
-                                <ExpansionPanelDetails>
-                                    <div>
-                                        <ReactMarkdown source={exp.content} escapeHtml={false} />
-                                    </div>
-                                </ExpansionPanelDetails>
-                            </ExpansionPanel>
-                        );
-                    })
-                }
+                    <ReactMarkdown source={project.content} escapeHtml={false} />
                 </div>
-            </ThemeProvider>
-        )
+                ),
+            }
+        })
     }
 
     render(){
         return (
             <TitleGridContainer title='Projects' desc={projectDescription}>
-            {
-                this.renderProjectMD(this.state.md)
-            }
+                <ListPanel panelList={this.transformProjectToPanels(this.state.md)} />
             </TitleGridContainer>
         )
     }
